ConstructorController now assigns a frame a name upon creating it
The name given to a frame is "Frame i = (x1, y1, z1) j = (x2, y2, z2)", where the x, y, and z values are replaced with the coordinates of that frame's endpoints. This ensures that every frame will be given a unique name, because we do not allow more than one frame to exist with the same pair of endpoints.

(Unique names are necessary for identifying frames in inter-process communication with SAP.)